+++ proto/control-api/src/lib.rs
@@ -31,6 +31,7 @@
     clippy::if_then_some_else_none,
     clippy::imprecise_flops,
     clippy::index_refutable_slice,
+    clippy::iter_with_drain,
     clippy::let_underscore_must_use,
     clippy::lossy_float_literal,
     clippy::map_err_ignore,
@@ -40,6 +41,7 @@
     clippy::multiple_inherent_impl,
     clippy::mutex_integer,
     clippy::nonstandard_macro_braces,
+    clippy::only_used_in_recursion,
     clippy::option_if_let_else,
     clippy::panic_in_result_fn,
     clippy::pedantic,
@@ -60,6 +62,7 @@
     clippy::todo,
     clippy::trailing_empty_array,
     clippy::trivial_regex,
+    clippy::try_err,
     clippy::undocumented_unsafe_blocks,
     clippy::unimplemented,
     clippy::unnecessary_self_imports,
@@ -89,5 +92,16 @@
     variant_size_differences
 )]
 
+pub mod callback;
+pub mod control;
 #[cfg(feature = "grpc")]
 pub mod grpc;
+
+#[doc(inline)]
+pub use self::{
+    callback::Api as CallbackApi,
+    control::{
+        endpoint, member, room, Api as ControlApi, Element, Elements, Endpoint,
+        Fid, Member, Ping, Pong, Room,
+    },
+};